verify broken images when they are displayed
Iterating all likes is not scalable. As broken images are displayed in the
web we report the fact that it's broken to the server and allow the server
to check if the image is in fact missing. If the server encounters an error
trying to get the image, we delete the corresponding record.